improve data cleaning on scraped rents
For anibis, offers which are not 'appartements' or 'maisons' are
dropped.
For tutti we do the same thing, and we also filter out sales offers.

merging does additional filtering before producing a single list of
offers.

Coordinate mapping is done only if the street number is present.
otherwise offers are dropped.